requirements.txt
numpy
pandas
matplotlib
requests
beautifulsoup4
cryptography
selenium
webdriver-manager
pyyaml
pymysql
func-timeout